changed tests, and added new tests
