--- ZigBeeValve/systemconfig.json
@@ -1,74 +1,0 @@
-{
-   "systemname": "testsys",
-   "unit_assets": [
-      {
-         "name": "Template",
-         "details": {
-            "Location": [
-               "Kitchen"
-            ]
-         },
-         "model": "",
-         "period": 10,
-         "setpoint": 20,
-         "APIkey": ""
-      }
-   ],
-   "services": [
-      {
-         "servicedefinition": "setpoint",
-         "details": {
-            "Forms": [
-               "SignalA_v1a"
-            ],
-            "Unit": [
-               "Celsius"
-            ]
-         },
-         "registrationPeriod": 0,
-         "costUnit": ""
-      }
-   ],
-   "protocolsNports": {
-      "coap": 0,
-      "http": 8870,
-      "https": 0
-   },
-   "distinguishedName": {
-      "Country": [
-         "SE"
-      ],
-      "Organization": [
-         "Luleaa University of Technology"
-      ],
-      "OrganizationalUnit": [
-         "CPS"
-      ],
-      "Locality": [
-         "Luleaa"
-      ],
-      "Province": [
-         "Norrbotten"
-      ],
-      "StreetAddress": null,
-      "PostalCode": null,
-      "SerialNumber": "",
-      "CommonName": "arrowhead.eu",
-      "Names": null,
-      "ExtraNames": null
-   },
-   "coreSystems": [
-      {
-         "coresystem": "serviceregistrar",
-         "url": "http://localhost:8443/serviceregistrar/registry"
-      },
-      {
-         "coresystem": "orchestrator",
-         "url": "http://localhost:8445/orchestrator/orchestration"
-      },
-      {
-         "coresystem": "ca",
-         "url": "http://localhost:9000/ca/certification"
-      }
-   ]
-}
